week08/requirements.txt
mediapipe==0.10.21
opencv-python
matplotlib
Pillow
protobuf
python-dotenv